# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a Sumwhere server API",
        "title": "Sumwhere API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://www.sumwhere.kr",
            "email": "qjadn0914@naver.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "auth.sumwhere.kr",
    "basePath": "/v1",
    "paths": {
        "/login": {
            "post": {
                "description": "로그인",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "signin\u0026signup"
                ],
                "summary": "default 로그인",
                "operationId": "get-string-by-int",
                "parameters": [
                    {
                        "description": "email,password만 사용",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/auth.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/auth.Token"
                        }
                    }
                }
            }
        },
        "/login/facebook": {
            "post": {
                "description": "페이스북으로 로그인",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "signin\u0026signup"
                ],
                "summary": "페이스북으로 로그인",
                "parameters": [
                    {
                        "description": "페이스북 토큰",
                        "name": "token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/auth.Token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/auth.Token"
                        }
                    }
                }
            }
        },
        "/login/kakao": {
            "post": {
                "description": "카카오로 로그인",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "signin\u0026signup"
                ],
                "summary": "카카오 로그인",
                "parameters": [
                    {
                        "description": "카카오 토큰",
                        "name": "token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/auth.Token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/auth.Token"
                        }
                    }
                }
            }
        },
        "/profile": {
            "post": {
                "description": "create user profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "프로필 생성",
                "operationId": "get-string-by-int",
                "parameters": [
                    {
                        "description": "profile",
                        "name": "profile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/profile.Profile"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/profile.Profile"
                        }
                    }
                }
            }
        },
        "/profile/city": {
            "get": {
                "description": "전국 시 모델 가져오기",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "전국 시 가져오기",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/profile.Area"
                            }
                        },
                        "headers": {
                            "Token": {
                                "type": "string",
                                "description": "qwerty"
                            }
                        }
                    }
                }
            }
        },
        "/profile/district": {
            "get": {
                "description": "전국 시 모델 가져오기",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "전국 시 모델 가져오기",
                "parameters": [
                    {
                        "type": "string",
                        "description": "원하는 도시",
                        "name": "city",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/profile.Area"
                            }
                        },
                        "headers": {
                            "Token": {
                                "type": "string",
                                "description": "qwerty"
                            }
                        }
                    }
                }
            }
        },
        "/profile/image/{id}": {
            "post": {
                "description": "set user profile image",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "프로필 이미지 추가",
                "operationId": "profile",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "image must 3",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        },
                        "headers": {
                            "Token": {
                                "type": "string",
                                "description": "qwerty"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/utils.ApiError"
                        }
                    }
                }
            }
        },
        "/profile/{id}": {
            "get": {
                "description": "get user profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "프로필 가져오기",
                "operationId": "get-string-by-int",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/profile.Profile"
                        },
                        "headers": {
                            "Token": {
                                "type": "string",
                                "description": "qwerty"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "update user profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "프로필 업데이트",
                "operationId": "get-string-by-int",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/profile.Profile"
                        },
                        "headers": {
                            "Token": {
                                "type": "string",
                                "description": "qwerty"
                            }
                        }
                    }
                }
            }
        },
        "/signup/email": {
            "post": {
                "description": "로그인",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "signin\u0026signup"
                ],
                "summary": "회원가입",
                "parameters": [
                    {
                        "description": "유저모델",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/auth.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/auth.Token"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.Token": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eapihjrphij32249j_erg94-2gojEGG32g2"
                }
            }
        },
        "auth.User": {
            "type": "object",
            "properties": {
                "admin": {
                    "type": "boolean"
                },
                "createAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "profile.Area": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string",
                    "example": "서울특별시"
                },
                "createdAt": {
                    "type": "string"
                },
                "district": {
                    "type": "string",
                    "example": "강남구"
                },
                "id": {
                    "type": "integer",
                    "example": 0
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "profile.Profile": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer",
                    "example": 30
                },
                "area": {
                    "type": "integer"
                },
                "createAt": {
                    "type": "string"
                },
                "gender": {
                    "type": "string",
                    "example": "남성"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "interest": {
                    "type": "integer"
                },
                "job": {
                    "type": "string",
                    "example": "공무원"
                },
                "phone": {
                    "type": "string",
                    "example": "01051416906"
                },
                "updateAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "utils.ApiError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}